Update acs_ingest.py
Small bug fix - The ACS ingest script was uploading more documents each loop.

On the first chunk, it uploaded the first chunk. On the second chunk, it uploaded the first, and the second. On the third, it uploaded the first, second, and third, and so on. This meant the code for uploading to the ACS index was taking exponentially longer than it was meant to.