Documentation: Moved the configuration and env instructions (#683)
I found the ordering of when to update the env, config, and data folder
to be confusing. In my opinion, they should come after the resources
have been provisioned. They also have to be done regardless of how you
decide to run the accelerator locally so having them in one of the
options and not the other was confusing.

Co-authored-by: Julia Meshcheryakova <juliame@microsoft.com>