Bug/llm context recall fixes (#867)
The function `llm_context_recall` was out of date with the method
signature for `response_generator.generate_response` and improperly
parsing the latest response from that method. This was causing it to
throw many errors when you attempt to use it as an eval metric.

This PR brings it back up-to-date with the method signature and response
parsing.